refactor(logcat): split query helpers
